made auth guard methods isAdmin and isCreator

+++ app/Components/Comment/Services/CommentService.php
@@ -3,16 +3,18 @@
 namespace App\Components\Comment\Services;
 
 use App\Components\Comment\Models\Comment;
+use App\Components\User\Models\User;
 use App\Exceptions\PermissionDeniedException;
 use Auth;
 
 class CommentService implements CommentServiceContract
 {
-    private $comment;
+    private $comment, $user;
 
-    public function __construct(Comment $comment)
+    public function __construct(Comment $comment, User $user)
     {
         $this->comment = $comment;
+        $this->user = $user;
     }
 
     public function store($data)
@@ -27,13 +29,21 @@ class CommentService implements CommentServiceContract
 
     public function update($data, $commentId)
     {
-        $this->comment->findOrFail($commentId)->update($data);
-        return $this->comment->findOrFail($commentId)->toArray();
+        if ($this->user->isAdministrator() || $this->user->isCreator($commentId)) {
+            $this->comment->findOrFail($commentId)->update($data);
+            return $this->comment->findOrFail($commentId)->toArray();
+        } else {
+            throw new PermissionDeniedException ('This action is not allowed for you!');
+        }
     }
 
     public function destroy($commentId)
     {
-        return ['success' => $this->comment->findOrFail($commentId)->delete()];
+        if ($this->user->isAdministrator() || $this->user->isCreator($commentId)) {
+            return ['success' => $this->comment->findOrFail($commentId)->delete()];
+        } else {
+            throw new PermissionDeniedException ('This action is not allowed for you!');
+        }
     }
 
     public function list($postId)
